chore(#395): address comments
